refactor(deployconform): consolidate source files along pipeline stages; document startup gating
Merge redact into canonical.go (canonical form: serialization + redaction),
verify into baseline.go (generation and verification are the two symmetric
halves of the conformance comparison), and the in-memory cache into
resolve.go (its only consumer). 11 source files become 8, one per pipeline
stage; no behavior change, tests untouched and passing.

README: add an optional compose-level startup gate
(depends_on: service_completed_successfully + verify --strict) with its
trust and availability caveats.
